+++ CHANGELOG.md
@@ -1,3 +1,21 @@
 # 0.0.1
 
-- feat: initial commit 🎉
+## Initial rRelease 🎉
+
+### Features:
+
+**Commands**
+
+  - `dfn`: core tool
+  - `dfn config`: for managing scripts
+  - `dfn config add <script|path>`: for registering scripts
+  - `dfn config remove <script|path>`: for un-registering scripts
+  - `dfn config rm <script|path>`: - _`dfn config remove <script|path>` alias_
+  - `dfn list`: for showing all registered scripts
+  - `dfn ls`: - _`dfn list` alias_
+
+**Options**
+
+  - `dfn --help`: show usage/help
+  - `dfn -h`: _`dfn --help` alias_
+  - `dfn --verbose`: enable verbose logging